requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.10.6
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Database drivers
pymysql==1.1.0
psycopg2-binary==2.9.9

# Additional utilities
slugify==0.0.1
python-slugify==8.0.1 